Split cursor name resolution out of PInvokeGenerator.cs
Move the contiguous cursor-name and remapping helpers (GetCursorName,
GetCursorQualifiedName, GetRemappedCursorName, GetRemappedName,
GetRemappedNameForAnonymousRecord, GetRemappedTypeName, and the adjacent
GetOverloadIndex/GetRecordDecl/GetExprAsWritten resolution helpers) into a
dedicated PInvokeGenerator.Naming.cs partial. The main partial is now
~3930 lines, down from ~7350. Pure code move; no behavior change.
